refactor: Kakao API 응답 DTO를 인프라 계층으로 이동
