+++ src/main/java/jeonseguard/backend/auth/application/facade/AuthFacade.java
@@ -1,6 +1,7 @@
 package jeonseguard.backend.auth.application.facade;
 
 import jeonseguard.backend.auth.application.service.AuthService;
+import jeonseguard.backend.auth.infrastructure.dto.KakaoUserInfoResponse;
 import jeonseguard.backend.auth.presentation.dto.request.*;
 import jeonseguard.backend.auth.presentation.dto.response.*;
 import jeonseguard.backend.user.application.service.UserService;
