+++ src/main/java/jeonseguard/backend/auth/application/service/AuthService.java
@@ -1,6 +1,7 @@
 package jeonseguard.backend.auth.application.service;
 
 import jeonseguard.backend.auth.domain.*;
+import jeonseguard.backend.auth.infrastructure.dto.*;
 import jeonseguard.backend.auth.infrastructure.provider.*;
 import jeonseguard.backend.auth.presentation.dto.request.*;
 import jeonseguard.backend.auth.presentation.dto.response.*;
