+++ src/main/java/jeonseguard/backend/auth/infrastructure/dto/KakaoTokenResponse.java
@@ -1,4 +1,4 @@
-package jeonseguard.backend.auth.presentation.dto.response;
+package jeonseguard.backend.auth.infrastructure.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.v3.oas.annotations.media.Schema;
